Resource Pool Submission Complete
The submit to resource pool functionality is complete. Going to work on the reward function for both resoruces and depleted islands.

+++ WarGame/Assets/Scripts/UnitTesting/Tests.cs
@@ -47,7 +47,7 @@ public class Tests : MonoBehaviour
 
         ResourceOrder order = new ResourceOrder();
         order.rsrc = 0;
-        order.amnt = new List<uint> { 0, 1000, 1000 };
+        order.amnt = new List<double> { 0, 1000, 1000 };
 
         List<string> depletedIslands = new List<string> { "0", "A3", "BB", "1" };
